+++ CHANGELOG.md
@@ -2,7 +2,19 @@
 
 All notable changes to the "nsql" extension will be documented in this file.
 
-## [0.0.1] - 2025-04-24
+## [0.2.0] - 2025-04-28
+
+### Added
+
+- Added `NSQL` language server with basic hover and completion support.
+- Added `NSQL` document validation.
+
+### Changed
+
+- Updated package.json to include necessary dependencies for the language server.
+- Updated .vscodeignore to exclude node_modules.
+
+## [0.1.0] - 2025-04-24
 
 ### Added
 
@@ -10,6 +22,11 @@ All notable changes to the "nsql" extension will be documented in this file.
 - LICENSE file
 - @vscode/vsce dev dependency
 - node_modules to the .gitignore
+
+### Changed
+
+- This version was previously released as 0.0.1, however, I have changed the versioning scheme to follow [Semantic Versioning](https://semver.org/) (MAJOR.MINOR.PATCH) to better reflect the changes made in each release.
+- The version number in package.json was changed from 0.0.1 to 0.1.0 to reflect the new versioning scheme.
 
 ## [Unreleased]